[Refactor] Align CSS parser numeric types
- Use size_t throughout tokenizer offset and range operations to match string sizes.
- Make clamped selector coefficient conversions explicit and avoid narrowing diagnostics.

issue: f-7350207575
AutoSubmit: true

+++ core/renderer/css/ng/parser/css_tokenizer_input_stream.cc
@@ -21,7 +21,7 @@ void CSSTokenizerInputStream::AdvanceUntilNonWhitespace() {
     ++offset_;
 }
 
-double CSSTokenizerInputStream::GetDouble(unsigned start, unsigned end) const {
+double CSSTokenizerInputStream::GetDouble(size_t start, size_t end) const {
   DCHECK(start <= end && ((offset_ + end) <= string_length_));
   bool is_result_ok = false;
   double result = 0.0;
